Fixing #8 by propogating errors.
Errors reading or writing sends an error to the dispatcher, from where
classes like monitor, pubsub and terminal exposes through their stream.
Where as client causes all pending and new commands to return an error.

Note. if consumer is not explicitely await the `Connection.done` an
Exception will occur here and crash the process, just as it currently
happens.